requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
xgboost>=1.5.0
catboost>=1.0.6
optuna>=3.0.0
pyyaml>=6.0
joblib>=1.1.0
kaggle>=1.5.12
